requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 데이터베이스 및 ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# 머신러닝 및 강화학습
torch==2.1.1
torchvision==0.16.1
stable-baselines3==2.2.1
gymnasium==0.29.1
numpy==1.24.4
scikit-learn==1.3.2
pandas==2.1.4

# 파일 모니터링
watchdog==3.0.0
aiofiles==23.2.1

# HTTP 클라이언트
httpx==0.25.2
aiohttp==3.9.1

# 로깅 및 모니터링
loguru==0.7.2
prometheus-client==0.19.0
structlog==23.2.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 개발 도구
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# 환경 변수
python-dotenv==1.0.0

# JSON 처리
orjson==3.9.10

# 시간 처리
python-dateutil==2.8.2

# 병렬 처리
asyncio-throttle==1.0.2

# 메시지 큐 (선택사항)
redis==5.0.1
celery==5.3.4

# API 문서
sphinx==7.2.6
sphinx-rtd-theme==1.3.0